refactor: 개선된 Ducks 패턴 적용
