+++ src/modules/chats/types.tsx
@@ -2,8 +2,8 @@ import { ActionType } from 'typesafe-actions';
 import { IChats } from '../../types';
 import * as actions from './actions';
 
-export type TAction = ActionType<typeof actions>;
+export type TChatsAction = ActionType<typeof actions>;
 
-export type TState = {
+export type TChatsState = {
   data: IChats;
 };
