+++ src/modules/receiver/actions.tsx
@@ -1,0 +1,7 @@
+import { createAction } from 'typesafe-actions';
+
+export const INITIALIZE = 'receiver/INITIALIZE';
+export const TOGGLE = 'receiver/TOGGLE';
+
+export const initReceiver = createAction(INITIALIZE)<number>();
+export const toggleReceiver = createAction(TOGGLE)<number>();
